Move release date to box
