+++ src/app/story/directives/status.component.ts
@@ -57,6 +57,21 @@ import { StoryModel } from '../../shared';
           <p *ngIf="!changed">Complete</p>
         </ng-container>
       </dd>
+      <dd>
+        <prx-fancy-field label="Release Date">
+          <div class="fancy-hint">
+            <input type="checkbox" [ngModel]="showReleasedAt" (click)="toggleShowReleaseAt()" name="showReleasedAt" id="showReleasedAt">
+            <label for="showReleasedAt">Specify date and time to be published</label>
+          </div>
+          <div class="fancy-hint" *ngIf="showReleasedAt">If you'd like to manually alter this episode's publication
+          to either delay or back-date its release, select the desired release date and time here.
+          Otherwise, the episode will be released immediately once published.
+          </div>
+          <prx-tz-datepicker *ngIf="showReleasedAt"
+            [date]="story.releasedAt" (dateChange)="story.set('releasedAt', $event)" [changed]="releasedAtChanged">
+          </prx-tz-datepicker>
+        </prx-fancy-field>
+      </dd>
 
     </dl>
   `
@@ -73,6 +88,7 @@ export class StoryStatusComponent implements DoCheck {
   isReleased: boolean;
   isScheduled: boolean;
   notPublished: boolean;
+  showReleasedAt: boolean;
 
   normalInvalid: string;
   normalInvalidCount: string;
@@ -99,6 +115,33 @@ export class StoryStatusComponent implements DoCheck {
       this.strictInvalid = this.storyInvalid(true);
       this.strictInvalidCount = this.countProblems(true);
       this.changed = this.story.changed();
+      if (this.story && this.story.releasedAt) {
+        this.showReleasedAt = true;
+      } else {
+        this.showReleasedAt = false
+      }
+    }
+  }
+
+  toggleShowReleaseAt() {
+    this.showReleasedAt = !this.showReleasedAt;
+    if (this.story.releasedAt) {
+      this.story.releasedAt = null;
+      this.notifyOfCanceledPublication();
+    } else {
+      this.story.releasedAt = new Date();
+    }
+  }
+
+  notifyOfCanceledPublication() {
+    const futurePublished = this.story.publishedAt && new Date() < this.story.publishedAt;
+    const removingReleaseDate = this.story.changed('releasedAt') && !this.story.releasedAt;
+    if (removingReleaseDate && futurePublished) {
+      this.modal.alert(
+        '',
+        'Removing the scheduled release date for a published episode will unpublish the episode.',
+        () => {}
+      );
     }
   }
